+++ src/xrGame/traffic_optimization.cpp
@@ -58,7 +58,7 @@ bool init_lzo(u8*& dest_wm, u8*& wm_buffer, lzo_dictionary_buffer& dest_dict)
     dest_dict.size = buffer_size;
 
     lzo_initialize();
-    wm_buffer = static_cast<u8*>(xr_malloc(lzo_get_workmem_size() + 16));
+    wm_buffer = static_cast<u8*>(xr_malloc(static_cast<size_t>(lzo_get_workmem_size()) + 16));
     // buffer must be alligned to 16 bytes
     dest_wm = (u8*)(size_t(wm_buffer + 16) & ~0xf);
     return true;
